Get College Basketball NET Rankings for the current date from the NCAA website

Get College Basketball NET Rankings for the current date from the NCAA website

cbb_ncaa_NET_rankings()

Value

Returns a tibble

Author

Saiem Gilani

Examples

# Get current NCAA NET rankings
cbb_ncaa_NET_rankings()

#> # A tibble: 347 x 12
#>     Rank Previous School Conference Record Road  Neutral Home  `Quad 1` `Quad 2`
#>    <int>    <int> <chr>  <chr>      <chr>  <chr> <chr>   <chr> <chr>    <chr>   
#>  1     1        1 Gonza~ WCC        22-0   7-0   5-0     10-0  7-0      5-0     
#>  2     2        2 Baylor Big 12     18-0   6-0   3-0     9-0   6-0      1-0     
#>  3     3        3 Michi~ Big Ten    16-1   5-1   0-0     11-0  6-1      4-0     
#>  4     4        4 Illin~ Big Ten    16-6   6-3   0-1     10-2  7-6      4-0     
#>  5     5        5 Iowa   Big Ten    17-6   5-3   0-1     12-2  5-5      6-1     
#>  6     6        6 Houst~ AAC        17-3   5-3   1-0     11-0  2-1      4-1     
#>  7     7        7 Ohio ~ Big Ten    18-5   7-3   1-0     10-2  8-4      4-1     
#>  8     8        8 Alaba~ SEC        18-5   5-2   2-2     11-1  6-3      6-2     
#>  9     9        9 Virgi~ ACC        15-5   5-3   1-2     9-0   3-4      5-0     
#> 10    10       10 Villa~ Big East   15-3   4-2   3-1     8-0   2-2      4-1     
#> # ... with 337 more rows, and 2 more variables: `Quad 3` <chr>, `Quad 4` <chr>